fix: no non-null judgement for `rm -rf`

Created-by: ChaosJohn
Commit-by: yujiahao
Merged-by: CangjiePL
Description: ## 变更内容（必填）

修复shell脚本的告警，`rm -rf` 未进行判空

## 变更类型（必填）

请描述本次Pull Request变更类型（原因），**请在保存后点击复选框或在编辑时将对应项前面的 `[ ]` 改为 `[X]` 。**

- [ ] 新增需求
- [ ] 问题修复
- [X] 构建过程或辅助工具变动
- [ ] 文档更新

## 变更内容自检（必填）

**请勿修改或删除以下选项内容，请在保存后点击复选框或在编辑时将对应项前面的 `[ ]` 改为 `[X]` 。**

### 编译本地自验证结果：

- [ ] 仓颉编译器编译通过
- [ ] 仓颉运行时编译通过
- [ ] 仓颉标准库编译通过
- [ ] 仓颉stdx库编译通过
- [X] 不涉及，辅助工具变动或者文档更新可以直接选择此项

### 测试用例本地自验证结果：

- [ ] 通过，请在下方提供截图证明
- [X] 不涉及，请在下方说明情况
shell脚本的告警修改，不涉及

### 文档示例代码编译运行结果：

- [ ] 通过，请在下方提供截图证明
- [X] 不涉及

## 关联的issue（必填）

https://gitcode.com/Cangjie/cangjie_stdx/issues/214


See merge request: Cangjie/cangjie_stdx!382

+++ downloader.sh
@@ -160,7 +160,9 @@ unzip -q "$CACHED_FILE" -d "$TMP_EXTRACT_DIR" || error_exit "Unzip failed. The f
 TOP_LEVEL_DIR=$(find "$TMP_EXTRACT_DIR" -mindepth 1 -maxdepth 1 -type d)
 
 if [ -z "$TOP_LEVEL_DIR" ] || [ ! -d "$TOP_LEVEL_DIR" ]; then
-    rm -rf "$TMP_EXTRACT_DIR"
+    if [ -d "$TMP_EXTRACT_DIR" ]; then
+        rm -rf "$TMP_EXTRACT_DIR"
+    fi
     error_exit "Could not find a top-level directory in the zip file."
 fi
 
@@ -173,7 +175,9 @@ fi
 mv "$TOP_LEVEL_DIR" "$FINAL_PATH"
 
 # Clean up temp directory
-rm -rf "$TMP_EXTRACT_DIR"
+if [ -d "$TMP_EXTRACT_DIR" ]; then
+    rm -rf "$TMP_EXTRACT_DIR"
+fi
 
 # 6. Post-extraction for macOS
 if [[ "$(uname -s)" == "Darwin" && "$PLATFORM_ARCH" == mac-* ]]; then
